frontend/course: react.fc the assignment header search bar
