+++ src/packages/frontend/course/common/consts.ts
@@ -1,0 +1,8 @@
+/*
+ *  This file is part of CoCalc: Copyright © 2020 Sagemath, Inc.
+ *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
+ */
+
+import { CSS } from "@cocalc/frontend/app-framework";
+
+export const SEARCH_STYLE: CSS = { marginBottom: "0px" } as const;
