+++ src/packages/frontend/course/common/index.ts
@@ -1,3 +1,8 @@
+/*
+ *  This file is part of CoCalc: Copyright © 2020 Sagemath, Inc.
+ *  License: AGPLv3 s.t. "Commons Clause" – see LICENSE.md for details
+ */
+
 export { FoldersToolbar } from "./folders-tool-bar";
 export { FlexPanel } from "./flex-panel";
 export { Progress } from "./progress";
